add client tests

+++ go.test.sh
@@ -3,10 +3,40 @@
 set -e
 echo "" > coverage.txt
 
-for d in $(go list ./... | grep -v vendor); do
-    go test -race -coverprofile=profile.out -covermode=atomic $d
+if [[ $TRAVIS_GO_VERSION == 1.6* ]]
+then
+
+  for d in $(go list ./... | grep -v vendor | grep -v http2 | grep -v example); do
+    go test -v -race -coverprofile=profile.out -covermode=atomic $d
     if [ -f profile.out ]; then
         cat profile.out >> coverage.txt
         rm profile.out
     fi
-done
+  done
+
+elif [[ $TRAVIS_GO_VERSION == 1.7* ]]
+then
+
+  for d in $(go list ./... | grep -v vendor | grep -v http2 | grep -v example); do
+    go test -v -race -coverprofile=profile.out -covermode=atomic $d
+    if [ -f profile.out ]; then
+        cat profile.out >> coverage.txt
+        rm profile.out
+    fi
+  done
+
+else
+
+  for d in $(go list ./... | grep -v vendor); do
+      go test -v -race $d
+  done
+
+  for d in $(go list ./... | grep -v vendor | grep -v http2 | grep -v example); do
+      go test -race -coverprofile=profile.out -covermode=atomic $d
+      if [ -f profile.out ]; then
+          cat profile.out >> coverage.txt
+          rm profile.out
+      fi
+  done
+
+fi
